requirements.txt
PILLOW~=9.4.0
pyperclip~=1.8.2